Add check for empty directory and version bump
This commit introduces a check in the index.js file if the targeted directory is empty. If true, it creates a new directory. This change ensures that a zip file is not generated for an empty directory. The version of the package has also been increased to 1.1.2 following this new addition to the functionality.